rename: templateHelpers -> repositoryURLs
